feat: implement DeleteCommunityInstanceService integrate lambda for community deletion

+++ app/workers/community_cleanup_worker.rb
@@ -6,7 +6,7 @@ class CommunityCleanupWorker
                            .where('deleted_at <= ?', 30.days.ago)
 
     if communities.empty?
-      Rails.logger.info "[CommunityCleanupWorker] No communities to clean up."
+      Rails.logger.info '[CommunityCleanupWorker] No communities to clean up.'
       return
     end
 
@@ -15,21 +15,29 @@ class CommunityCleanupWorker
     communities.find_each do |community|
       begin
         ActiveRecord::Base.transaction do
-          account_ids = community&.community_admins.pluck(:account_id)
+          account_ids = community&.community_admins&.pluck(:account_id)
 
           Rails.logger.info "[CommunityCleanupWorker] Deleting community ##{community.id}..."
-          community.destroy
 
-          account_ids.compact.uniq.each do |account_id|
-            if Account.exists?(account_id)
-              Admin::AccountDeletionWorker.perform_async(account_id, {
-                'reserve_username' => false,
-                'reserve_email' => false
-              })
-              Rails.logger.info "[CommunityCleanupWorker] Enqueued deletion for account ##{account_id}."
-            else
-              Rails.logger.warn "[CommunityCleanupWorker] Account ##{account_id} not found. Skipping deletion."
+          response = Patchwork::DeleteCommunityInstanceService.new.call(community.id)
+
+          if response
+            Rails.logger.info "[CommunityCleanupWorker] Successfully called DeleteCommunityInstanceService for community ##{community.id}."
+            community.destroy
+
+            account_ids.compact.uniq.each do |account_id|
+              if Account.exists?(account_id)
+                Admin::AccountDeletionWorker.perform_async(account_id, {
+                  'reserve_username' => false,
+                  'reserve_email' => false,
+                })
+                Rails.logger.info "[CommunityCleanupWorker] Enqueued deletion for account ##{account_id}."
+              else
+                Rails.logger.warn "[CommunityCleanupWorker] Account ##{account_id} not found. Skipping deletion."
+              end
             end
+          else
+            Rails.logger.warn "[CommunityCleanupWorker] Skipping destroy for community ##{community.id} due to service call failure."
           end
           sleep(0.05)
         end
@@ -38,6 +46,6 @@ class CommunityCleanupWorker
       end
     end
 
-    Rails.logger.info "[CommunityCleanupWorker] Cleanup completed."
+    Rails.logger.info '[CommunityCleanupWorker] Cleanup completed.'
   end
 end
